pyproject.toml
[project]
name = "dentaldesk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.18.3",
    "annotated-types==0.7.0",
    "asyncpg>=0.29.0",
    "bcrypt>=4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "anyio==4.12.1",
    "click==8.3.1",
    "fastapi==0.115.6",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "packaging==26.0",
    "pluggy==1.6.0",
    "pydantic[email]==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "redis>=7.1.1",
    "ruff==0.14.14",
    "sqlalchemy==2.0.46",
    "starlette==0.41.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.34.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "sqlmodel>=0.0.33",
    "psycopg[binary]>=3.3.2",
    "psycopg2>=2.9.11",
    "greenlet>=3.3.1",
]
